分页参数数据类型统一

+++ kcwl-ddd-core/src/main/java/com/kcwl/ddd/interfaces/query/PageQuery.java
@@ -4,10 +4,10 @@ package com.kcwl.ddd.interfaces.query;
 import com.kcwl.ddd.domain.entity.PageCondition;
 
 public class PageQuery extends BaseQuery {
-    private long curPagerNo = PageCondition.FIRST_PAGE_NO;
-    private long pageSize = PageCondition.DEFAULT_PAGE_SIZE;
+    private int curPagerNo = PageCondition.FIRST_PAGE_NO;
+    private int pageSize = PageCondition.DEFAULT_PAGE_SIZE;
 
-    public long getCurPagerNo() {
+    public int getCurPagerNo() {
         return curPagerNo;
     }
 
@@ -15,7 +15,7 @@ public class PageQuery extends BaseQuery {
         this.curPagerNo = curPagerNo;
     }
 
-    public long getPageSize() {
+    public int getPageSize() {
         return pageSize;
     }
 
@@ -23,7 +23,7 @@ public class PageQuery extends BaseQuery {
         this.pageSize = pageSize;
     }
 
-    public long getOffset() {
+    public int getOffset() {
         return (getCurPagerNo() - 1) * getPageSize();
     }
 }
